refactor(router): organize worker selection API
